Reforming the engine again.
The idea this time is that every entity should check for it's own whether
it is colliding with something and decide how to handle that collision in
terms of movement. The TileEnvironment will still handle the collision
resolving between entities and tiles. The collision checking that every
entity has to do has to do with deciding what animation should be played
and what velocity should be applied.

This couldn't be done in the old way because certain properties (like
walkRightAnimation) are only present in certain instances of entities.
Therefore, TileEngine couldn't reliably handle these cases.