Changed a lot.
I added the basics to use Entity Components and also added a
GravityComponent. It doesn't work correctly yet though, because
it keeps saying that it's owner's isOnGround property is true,
even when it really is false. I think it has something to do with
inheritance, because that property is defined in Entity, and the
error occurs in Player... The Player itself says the value is false
while it's GravityComponent says its true...